requirements.txt
Django
psycopg
django-cors-headers
djangorestframework
markdown
django-filter
graphene-django
pandas
selenium
beautifulsoup4
django-extensions
requests
lxml
html5lib
openai
sentence-transformers==1.2.0
umap-learn[plot]==0.5.1
hdbscan==0.8.26
nltk
contextualized-topic-models
torch
torchvision
scikit-learn
langchain[all]
